Network LWR Model update Split program : define_map is the input file for the network and junction_info dictionaries
