+++ density_plot.m
@@ -31,22 +31,14 @@ for i=1:338
         ylim([0 113])
         title('Road 1')
         text(5,100,sprintf('Frame %i',i))
-    subplot(2,4,2)
-        plot(1:25,road2(i,:),'o')
+    subplot(2,4,[2 3])
+        plot(1:50,[road2(i,:),road4(i,:)],'o')
         ylim([0 113])
-        title('Road 2')
-    subplot(2,4,6)
-        plot(1:25,road3(i,:),'o')
+        title('Road 2 and 4')
+    subplot(2,4,[6 7])
+        plot(1:50,[road3(i,:),road5(i,:)],'o')
         ylim([0 113])
-        title('Road 3')
-    subplot(2,4,3)
-        plot(1:25,road4(i,:),'o')
-        ylim([0 113])
-        title('Road 4')
-    subplot(2,4,7)
-        plot(1:25,road5(i,:),'o')
-        ylim([0 113])
-        title('Road 5')
+        title('Road 3 and 5')
     subplot(2,4,[4 8])
         plot(1:25,road6(i,:),'o')
         ylim([0 113])
